Help: Verbatim (in example only) -> working Pre.
The example Help outline had a style called Verbatim, which didn't do preformatted whitespace.  The script had support for a style called Pre, not in the outline, which did preformatted whitespace poorly.  Expand Verbatim into Pre and specify the whitespace in CSS, which simplifies the script and improves whitespace handling.