Strip non-English localizations when building for installation too.
In this case, the target build directory (which must be shared) contains symlinks to the frameworks, which we can modify prior to their being copied into Shroud’s bundle.

Ideally these should be in UninstalledProducts (and Sparkle does this), but things are currently broken and they seem to build into a directory that looks a bit like the bundle-relative one.  Will fix.